job-board/django-backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-rest-passwordreset>=1.2.1
django-cors-headers>=4.3.0
django-ratelimit>=4.0.0
django-redis>=5.3.0
Pillow>=10.0.0
